feat(generators): Along with an upgrade to Angular 1.5, we introduce some changes to our sub-generator names.
- Before:
```yo nodes:component``` - this used to create a directive, module, html, scss file in the common directory.

- Now:
```yo nodes:component``` - this will create a module + Angular 1.5 .component() and gives you the options of adding html, scss and defining the component type (module, component)

- New:
```yo nodes:ui-directive``` - this is the replacement for the old yo nodes:component. This creates a directive, module, html and scss file in the common directory
```yo nodes:ui-component``` - this creates an angular .component(), module, html and scss in the common directory.

This release should not break any previous code, as this release only changes the CLI commands.